ai_agents/requirements/orchestrator_requirements.txt
boto3>=1.34.0
bedrock-agentcore
strands-agents
python-dotenv
